pyproject.toml
[project]
name = "spotify-data-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "uvloop>=0.22.1",
    "httptools>=0.7.1",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
]

[tool.vercel.scripts]
build = "python main.py"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
    "httpx>=0.27.0",
]
